Separate artist and song classes into separate files, also add __init__
